Move `CustomTileProviderRepository` to :core:data
